create change weight table in reports

+++ run_create_report.py
@@ -2,7 +2,7 @@ import asyncio
 import logging
 from time import perf_counter
 
-from config_app import SLEEP
+from config_app import SLEEP, WIDTH_TABLES, DELETE_SCREENSHOTS
 from move_shot_to_word import create_formatted_doc
 from parce_screenshots import run_create_report
 
@@ -11,17 +11,19 @@ from utils import sleep_system, delete_auth_state, delete_screenshots
 if __name__ == "__main__":
     t1 = perf_counter()
     try:
-        delete_screenshots()
+        if DELETE_SCREENSHOTS:
+            delete_screenshots()
         delete_auth_state()
 
         asyncio.run(run_create_report())
-        create_formatted_doc(target_image_width_px=900)
+        create_formatted_doc(target_image_width_px=WIDTH_TABLES)
         print(f"{'*' * 100} \nElapsed: {perf_counter() - t1:.1f}s\n {'*' * 100}")
         logging.info("*" * 100)
         if SLEEP:
             sleep_system()
     finally:
-        delete_screenshots()
+        if DELETE_SCREENSHOTS:
+            delete_screenshots()
         # Удаляем, так как ТХ ПРО не видит регистрацию без проходки в регистрации
         try:
             delete_auth_state()
